requirements.txt
botasaurus
python-dotenv
requests